[wasm][debugger][tests] Extract Delegate tests to a new file
